backend/requirements.txt
yfinance
pandas
numpy
requests
fastapi
uvicorn[standard]
akshare
python-dotenv
futu-api
